コピペミス修正

+++ TVRemoteOmxplayer.py
@@ -86,9 +86,10 @@ class TVRemoteOmxplayer(CecClient):
   
   def __init__(self):
     super().__init__()
-  
   def play(self, url, *options):
-    cmd = f"omxplayer %s '{url}' " % ' '.join(options)
+    cmd = f"omxplayer '{url}' "
+    if len(options) > 0 :
+      cmd = cmd + ' '.join(options)
     print(cmd)
     cmd = shlex.split(cmd)
     self.p = subprocess.Popen(cmd,stdin=subprocess.PIPE,stdout=subprocess.PIPE)
@@ -97,13 +98,16 @@ class TVRemoteOmxplayer(CecClient):
     key = self.keymap_of_omxplayer[name_of_command]
     self.p.stdin.write(key)
     self.p.stdin.flush()
-  
+
   def exit(self):
     self.on_exit()
-  
+
   def on_exit(self):
-    self.p.terminate()
-    self.cec_proc.terminate()
+    if hasattr(self,'p') :
+      self.p.terminate()
+    if hasattr(self,'cec_proc') :
+      self.cec_proc.terminate()
+      
   
   def dispatch(self,key):
